fixed bug from prev commit

+++ com.reflexit.magiccards.core/src/com/reflexit/magiccards/core/sync/CardCache.java
@@ -191,9 +191,11 @@ public class CardCache {
 			FileUtils.saveStream(st, file2);
 			st.close();
 			if (file2.exists()) {
-				boolean rem = file.delete();
-				if (rem == false)
-					throw new IOException("failed to delete " + file.toString());
+				if (file.exists()) {
+					boolean rem = file.delete();
+					if (rem == false)
+						throw new IOException("failed to delete " + file.toString());
+				}
 				boolean did = file2.renameTo(file);
 				if (!file.exists() || did == false) {
 					throw new IOException("failed to rename into " + file.toString());
